Add cap to hydro reservoir time series
The hydro reservoir "capacity factor" inflow time series has values as high as 1e2 on the regional level. Together with minimal capacity factors of 1e-3 this leads to a numerical range of 1e6. For the numerics of barrier, this can be a problem. To reduce this to 1e5, I'm now capping the reservoir time series at 10. That is, the highest inflow is 10 times the installed capacity. Less than 1% of all time steps are affected. The total annual inflow stays the same.