Add small variable costs for wind solar
Variable costs are very small as we only introduce them so that curtailment is enforced and electricity is not burned on transmission lines. The costs should not be too small as otherwise the numerical range of the problem would suffer too much and barrier may have problems finding solutions. Thus we go for 0.1 ct/kWh.

Because 0.1 ct/kWh are a significant increase of costs (~5% for solar PV) we deduct those costs from fixed O&M costs based on estimated yearly generation.